# optimize `float_parse_bytes`