Noticed some opportunity while reading #81, which will help mitigate some of the new cost on floats.

Before:

```
running 1 test
test parse_timestamp_str             ... bench:         375.03 ns/iter (+/- 15.90)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured; 18 filtered out; finished in 0.29s
```

After:

```
running 1 test
test parse_timestamp_str             ... bench:         327.23 ns/iter (+/- 19.36)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured; 18 filtered out; finished in 0.22s
```