@davidhewitt 2025-02-11
Good idea, refactored to also cover integer parsing. This seems to have worked out even better:

```
running 1 test
test parse_timestamp_str             ... bench:         319.49 ns/iter (+/- 12.95)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured; 18 filtered out; finished in 2.04s
```